+++ tests/integration/testdata/buildcmd/CDK/cdk-example-layer/requirements.txt
@@ -1,1 +1,3 @@
--e .
+aws-cdk.core
+aws-cdk.aws_lambda
+aws-cdk.aws_apigateway
